fix: Fold the noise seed into the deterministic noiser
DeterministicTruncatedLaplaceNoise held a single function after the
draws moved into ResultNoiser. Folds fingerprint into
DeterministicTruncatedLaplaceResultNoiser, which now takes the combined
frequency vector and contribution count and derives the seed itself, so
no caller assembles it.

Also fixes the GaussianResultNoiser call sites in the direct result
builders, which the previous commit left on the one-argument form.

Issue: #4183
